refactor : ErrorCode global 패키지로 수정 및 excpetion처리 리팩토링
