+++ src/main/java/com/codeit/duckhu/domain/notification/exception/NotificationAccessDeniedException.java
@@ -1,15 +1,12 @@
 package com.codeit.duckhu.domain.notification.exception;
 
+import com.codeit.duckhu.global.exception.ErrorCode;
 import java.util.Map;
 import java.util.UUID;
 
 public class NotificationAccessDeniedException extends NotificationException {
 
-  public NotificationAccessDeniedException(UUID userId, UUID notificationId) {
-    super(
-        NotificationErrorCode.INVALID_NOTIFICATION_RECEIVER,
-        Map.of(
-            "userId", userId.toString(),
-            "notificationId", notificationId.toString()));
+  public NotificationAccessDeniedException(ErrorCode errorCode) {
+    super(errorCode);
   }
 }
