+++ src/main/java/com/codeit/duckhu/domain/notification/exception/NotificationAlreadyConfirmedException.java
@@ -1,0 +1,9 @@
+package com.codeit.duckhu.domain.notification.exception;
+
+import com.codeit.duckhu.global.exception.ErrorCode;
+
+public class NotificationAlreadyConfirmedException extends NotificationException {
+  public NotificationAlreadyConfirmedException(ErrorCode errorCode) {
+    super(errorCode);
+  }
+}
