+++ src/main/java/com/codeit/duckhu/domain/notification/exception/NotificationNotFoundException.java
@@ -1,13 +1,12 @@
 package com.codeit.duckhu.domain.notification.exception;
 
+import com.codeit.duckhu.global.exception.ErrorCode;
 import java.util.Map;
 import java.util.UUID;
 
 public class NotificationNotFoundException extends NotificationException {
 
-  public NotificationNotFoundException(UUID notificationId) {
-    super(
-        NotificationErrorCode.NOTIFICATION_NOT_FOUND,
-        Map.of("notificationId", notificationId.toString()));
+  public NotificationNotFoundException(ErrorCode errorCode) {
+      super(errorCode);
   }
 }
