+++ src/main/resources/application.yml
@@ -14,7 +14,10 @@ spring:
     database: postgresql
   transaction:
     default-timeout: 30
-
+  springdoc:
+    swagger-ui:
+      enabled: true
+      path: /swagger-custom-ui.html
 duckhu:
   storage:
     type: s3
